requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.txt
#
aniso8601==9.0.1
    # via
    #   -r requirements.txt
    #   flask-restful
click==8.0.3
    # via
    #   -r requirements.txt
    #   flask
dnspython==2.1.0
    # via
    #   -r requirements.txt
    #   email-validator
email-validator==1.1.3
    # via
    #   -r requirements.txt
    #   wtforms
flask==2.0.2
    # via
    #   -r requirements.txt
    #   flask-mongoengine
    #   flask-restful
    #   flask-wtf
flask-mongoengine==1.0.0
    # via -r requirements.txt
flask-restful==0.3.9
    # via -r requirements.txt
flask-wtf==1.0.0
    # via
    #   -r requirements.txt
    #   flask-mongoengine
idna==3.3
    # via
    #   -r requirements.txt
    #   email-validator
itsdangerous==2.0.1
    # via
    #   -r requirements.txt
    #   flask
    #   flask-wtf
jinja2==3.0.3
    # via
    #   -r requirements.txt
    #   flask
markupsafe==2.0.1
    # via
    #   -r requirements.txt
    #   jinja2
    #   wtforms
marshmallow==3.14.1
    # via marshmallow-mongoengine
marshmallow-mongoengine==0.30.2
    # via -r requirements.txt
mongoengine==0.23.1
    # via
    #   -r requirements.txt
    #   flask-mongoengine
    #   marshmallow-mongoengine
pymongo==3.12.1
    # via
    #   -r requirements.txt
    #   mongoengine
python-dotenv==0.19.2
    # via -r requirements.txt
pytz==2021.3
    # via
    #   -r requirements.txt
    #   flask-restful
six==1.16.0
    # via
    #   -r requirements.txt
    #   flask-restful
werkzeug==2.0.2
    # via
    #   -r requirements.txt
    #   flask
wtforms[email]==3.0.0
    # via
    #   -r requirements.txt
    #   flask-mongoengine
    #   flask-wtf
